Move MockSubscriber to mocks dir
